Move torchax dflash model and spec_decode wrapper to vllm
